+++ shapes/cube.ts
