+++ shapes/icosphere_flat.blend
