+++ shapes/icosphere_smooth.blend
