+++ shapes/icosphere_smooth.ts
@@ -1,7 +1,7 @@
 import {Shape} from "../common/material.js";
 import {GL_ARRAY_BUFFER, GL_ELEMENT_ARRAY_BUFFER, GL_STATIC_DRAW} from "../common/webgl.js";
 
-export function mesh_IcosphereSmooth(gl: WebGL2RenderingContext) {
+export function mesh_icosphere_smooth(gl: WebGL2RenderingContext) {
     let Vertices = gl.createBuffer();
     gl.bindBuffer(GL_ARRAY_BUFFER, Vertices);
     gl.bufferData(GL_ARRAY_BUFFER, vertices, GL_STATIC_DRAW);
