+++ shapes/monkey_flat.blend
