+++ shapes/monkey_flat.ts
@@ -1,7 +1,7 @@
 import {Shape} from "../common/material.js";
 import {GL_ARRAY_BUFFER, GL_ELEMENT_ARRAY_BUFFER, GL_STATIC_DRAW} from "../common/webgl.js";
 
-export function mesh_MonkeyFlat(gl: WebGL2RenderingContext) {
+export function mesh_monkey_flat(gl: WebGL2RenderingContext) {
     let Vertices = gl.createBuffer();
     gl.bindBuffer(GL_ARRAY_BUFFER, Vertices);
     gl.bufferData(GL_ARRAY_BUFFER, vertices, GL_STATIC_DRAW);
