+++ shapes/monkey_smooth.blend
